requirements-test.txt
# ORION Test Dependencies
#
# Install with: pip install -r requirements-test.txt
#
# These are ONLY for testing, not for runtime.

# Testing framework
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# JSON Schema validation
jsonschema==4.21.1

# YAML parsing (for policy tests)
pyyaml==6.0.1

# Mocking external dependencies
fakeredis==2.21.1
pytest-asyncio==0.21.1
